changed decimal to double for SQLite compatible

+++ MovieList/DTO/MovieDTO.cs
@@ -10,16 +10,16 @@ public class MovieDTO
     
     public string? Title { get; set; }
     
-    public decimal? Popularity { get; set; }
+    public double? Popularity { get; set; }
     
     [DefaultValue(null)]
     public DateTime? ReleaseDate { get; set; } 
     
-    public decimal? Revenue { get; set; }
+    public double? Revenue { get; set; }
     
     public string? Status { get; set; }
     
-    public decimal? VoteAverage { get; set; }
+    public double? VoteAverage { get; set; }
     
-    public decimal? VoteCount { get; set; }
+    public double? VoteCount { get; set; }
 }
